pyproject.toml
[tool.poetry]
name = "ai-optimize-wealth-strategist"
version = "0.1.0"
description = "An AI-powered wealth management system that uses multiple agents for personalized portfolio management and financial planning"
authors = ["Optimize Asset Management <info@optimizeam.com>"]
readme = "README.md"
packages = [
    { include = "src", from = "." }
]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "0.3.0"
langchain-anthropic = "0.3.5"
langchain-groq = "0.2.3"
langchain-openai = "^0.3.5"
langchain-deepseek = "^0.1.2"
langchain-ollama = "^0.2.0"
langgraph = "0.2.56"
pandas = "^2.1.0"
numpy = "^1.24.0"
python-dotenv = "1.0.0"
matplotlib = "^3.9.2"
networkx = "^3.2.0"
tabulate = "^0.9.0"
colorama = "^0.4.6"
questionary = "^2.1.0"
rich = "^13.9.4"
langchain-google-genai = "^2.0.11"
# Backend dependencies
fastapi = {extras = ["standard"], version = "^0.104.0"}
fastapi-cli = "^0.0.7"
pydantic = "^2.4.2"
httpx = "^0.27.0"
sqlalchemy = "^2.0.22"
alembic = "^1.12.0"
# Financial data and analysis
yfinance = "^0.2.18"
websockets = "^14.0"
requests = "^2.31.0"
python-dateutil = "^2.8.2"

# Market data APIs (Phase 2 - Additional sources)
polygon-api-client = "^1.12.0"  # Real-time US equities, forex, crypto
# alpha-vantage = "^2.3.1"        # Technical indicators, Canadian stocks
newsapi-python = "^0.2.7"       # Financial news and sentiment
finnhub-python = "^2.4.18"      # Real-time financial news (Phase 3)
fredapi = "^0.5.1"              # US economic indicators

# Market data APIs (Phase 3 - Additional sources)
# Using requests from above for Phase 3 APIs
# twelve-data = "^1.0.0"
# marketstack = "^1.0.0"

# Canadian financial context
pytz = "^2023.3"
# Portfolio optimization
scipy = "^1.11.0"
scikit-learn = "^1.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 420
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
force_alphabetical_sort_within_sections = true 